backend/requirements.custom.txt
# Shared
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
ruff
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Custom pipeline GPU
pynvvideocodec==2.1.0
cupy-cuda12x>=13.0.0

# Vendored BOTSORT dependencies (no ultralytics/torch needed)
lap>=0.5.12
scipy>=1.14.0
Pillow>=10.0.0
